added validation (Validator interfaces&Impl) added specific exceptions added tests for above functionality

+++ src/main/java/lv/javaguru/java2/core/GenericServiceImpl.java
@@ -10,13 +10,13 @@ public abstract class GenericServiceImpl
        < T extends GenericDAO,
         R extends Generic> implements  GenericService {
     @Autowired
-    private T dao;
+    public T dao;
 
-    public List<Generic> getAll()  {
+    public List<Generic> getAll() throws Exception  {
         return dao.getAll();
     }
 
-    public Generic getObject(long id) {
+    public Generic getObject(long id) throws Exception{
         R result=null;
         try {
             result = (R)dao.getById(id);
@@ -31,20 +31,28 @@ public abstract class GenericServiceImpl
     }
 
     public void addObject(Generic object) throws Exception {
-        validate(object);
+        validate(object, true);
         dao.create(object);
     }
 
     public void updateObject(Generic object) throws Exception {
-        validate(object);
+        R result = (R)dao.getById(object.getId());
+        if (result==null)
+            throw new ObjectNotExistException();
+
+        validate(object, false);
         dao.update(object);
     }
 
-    public void deleteObject(long id)  {
+    public void deleteObject(long id) throws Exception {
+        R result = (R)dao.getById(id);
+        if (result==null)
+            throw new ObjectNotExistException();
+
         dao.delete(id);
     }
 
-    protected abstract void validate(Generic object) throws Exception;
+    public abstract void validate(Generic object, boolean add) throws Exception;
 
     public abstract R getNewInstance();
 }
